remove how2

+++ install.sh
@@ -2,10 +2,10 @@
 
 set -xe
 
-if type apt-get &> /dev/null; then
-	PM=apt-get
-	sudo apt-get update
-	sudo apt-get install -yyq software-properties-common build-essential cmake stow mpv vim
+if type apt &> /dev/null; then
+	PM=apt
+	sudo apt update
+	sudo apt install -yyq software-properties-common build-essential cmake stow mpv vim
 elif type yum &> /dev/null; then
 	PM=yum
 	sudo yum update
@@ -19,7 +19,7 @@ PATH_DIR="${HOME}/.local/bin"
 [ -d "$PATH_DIR" ] || mkdir -p "$PATH_DIR"
 
 # fish
-which fish >/dev/null || { sudo apt-add-repository ppa:fish-shell/release-3; sudo apt update; sudo apt install fish; }
+command -v fish >/dev/null || { sudo apt-add-repository ppa:fish-shell/release-3; sudo apt update; sudo apt install fish; }
 
 # cheat.sh
 curl https://cht.sh/:cht.sh > "$PATH_DIR/cht.sh"
@@ -27,7 +27,7 @@ chmod +x "$PATH_DIR/cht.sh"
 
 # batcat
 sudo apt install bat &&
-ln -s /usr/bin/batcat "$PATH_DIR/bat"
+[ -f "$PATH_DIR/bat" ] || ln -s /usr/bin/batcat "$PATH_DIR/bat"
 	
 # exa
 curl -s https://api.github.com/repos/ogham/exa/releases/latest | grep browser_download_url | grep linux-$(uname -m)-v | cut -d '"' -f 4 | wget -i- -qO- | busybox unzip - 
@@ -39,14 +39,11 @@ chmod +x "$PATH_DIR/exa"
 sudo apt install fd-find
 
 # fzf
-sudo apt-get install fzf
+sudo apt install fzf
 
 # has
 curl -sL https://git.io/_has | tee "$PATH_DIR/has" >/dev/null
 chmod +x "$PATH_DIR/has"
-
-# how2
-which how2 >/dev/null || npm install -g how2
 
 # tldr
 curl -o "$PATH_DIR/tldr" https://raw.githubusercontent.com/raylee/tldr/master/tldr &&
